% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{apply_grid_penalty}
\alias{apply_grid_penalty}
\title{Apply a Grid Penalty to a Driver}
\usage{
apply_grid_penalty(race_data = generate_next_race_data(), driver_id, penalty)
}
\arguments{
\item{race_data}{A data frame containing race participants. Must include
`driver_id`, `quali_position` and `grid`.}

\item{driver_id}{The character string ID of the driver to penalize.}

\item{penalty}{A positive integer representing the number of grid places
for the penalty.}
}
\value{
A tibble identical to `race_data` but with the `grid` column
  updated to reflect the penalty.
}
\description{
Adjusts the starting grid by applying a penalty to a specific driver. The
penalized driver is moved down the grid, and other drivers are moved up to
fill the vacated spot.
}
\details{
The function uses the `quali_position` column as the initial grid order.
It calculates the new position for the penalized driver and re-orders the
grid accordingly. The penalty is capped at the last position on the grid
(e.g., a 5-place penalty for a driver in 19th out of 20 results in a 20th
place start).
}
\examples{
\dontrun{
  # Assume `after_quali_data` is a data frame with quali results
  # Apply a 5-place penalty to 'max_verstappen'
  new_grid_data <- apply_grid_penalty(
    race_data = generate_next_race_data(),
    driver_id = "max_verstappen",
    penalty = 5
  )
}
}
