% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{default parameters
Just a list of default parameters}
\format{
An object of class \code{list} of length 12.
}
\usage{
default_params
}
\description{
default parameters
Just a list of default parameters
}
\keyword{internal}
