% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_quali_pos}
\alias{.predict_quali_pos}
\title{Predict Qualifying Position (Internal)}
\usage{
.predict_quali_pos(new_data, quali_pos_model)
}
\arguments{
\item{new_data}{A data frame of new data.}

\item{quali_pos_model}{A `workflow` or `model_stack` object for qualifying
position prediction.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and
  `likely_quali_position`.
}
\description{
Predict Qualifying Position (Internal)
}
\keyword{internal}
