% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_t10}
\alias{.predict_t10}
\title{Predict Top 10 Finish}
\usage{
.predict_t10(new_data = generate_next_race_data(), t10_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{t10_model}{A `workflow` object for predicting a top 10 finish.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `t10_odd`.
}
\description{
Predict Top 10 Finish
}
\keyword{internal}
