% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{.predict_winner}
\alias{.predict_winner}
\title{Predict Race Winner}
\usage{
.predict_winner(new_data = generate_next_race_data(), win_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{win_model}{A `workflow` object for predicting the winner.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `win_odd`.
}
\description{
Predict Race Winner
}
\keyword{internal}
