% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{format_quali_skeet_predictions}
\alias{format_quali_skeet_predictions}
\title{Format Qualifying Predictions for a Skeet}
\usage{
format_quali_skeet_predictions(predictions)
}
\arguments{
\item{predictions}{A data frame of predictions from `predict_quali_round()`}
}
\value{
A list containing the formatted string for the skeet body and a vector of tags.
}
\description{
Format Qualifying Predictions for a Skeet
}
