% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{format_race_skeet_odds}
\alias{format_race_skeet_odds}
\title{Format Race Odds for a Skeet}
\usage{
format_race_skeet_odds(predictions)
}
\arguments{
\item{predictions}{A data frame of predictions from `predict_round()`}
}
\value{
A list containing the formatted string for the skeet body and a vector of tags.
}
\description{
Format Race Odds for a Skeet
}
