% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{format_tweet_predictions}
\alias{format_tweet_predictions}
\title{Format Predictions for a Tweet}
\usage{
format_tweet_predictions(predictions)
}
\arguments{
\item{predictions}{A data frame of predictions from `predict_round()`}
}
\value{
A formatted string ready for tweeting.
}
\description{
Format Predictions for a Tweet
}
