% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{generate_new_data}
\alias{generate_new_data}
\title{Generate New Data for Prediction}
\usage{
generate_new_data(
  season,
  round,
  drivers = NULL,
  penalties = NULL,
  historical_data = clean_data(),
  params = get_processing_params(),
  use_live_data = TRUE
)
}
\arguments{
\item{season}{A numeric value for the season year (e.g., 2023).}

\item{round}{A numeric value for the round number of the season.}

\item{drivers}{A data frame or tibble containing the drivers for the race.
It must include `driver_id` and `constructor_id`. Optionally, it can
include `quali_position`, `grid`, and practice-related ranks. If `NULL`
(the default), it will use the drivers from the most recent race.}

\item{penalties}{A named vector of `driver_id = penalty_positions` to be
applied (in order of application).
For example `c('hamilton' = 5, 'max_verstappen' = 10)`.}

\item{historical_data}{A data frame of historical race data, typically the
output of `clean_data()`. This is used to calculate the feature values.}

\item{params}{A named list of processing parameters controlling imputation
defaults. Defaults to `get_processing_params()`.}

\item{use_live_data}{Whether or not to check for weekend data (e.g. recent qualis).
Default TRUE}
}
\value{
A tibble where each row corresponds to a driver for the specified
  race, and columns are the features required for the modeling functions.
}
\description{
Creates a feature set for a specific upcoming race for a given set of drivers.
This function calculates various driver, constructor, and circuit-specific
features based on historical data. It's designed to generate the input data
required by the prediction and simulation functions (e.g., [simulate_race()]).
}
\details{
The function takes a season, round, and a list of drivers to generate a
predictive dataset. It pulls historical performance data for each driver,
their constructor, and the specific circuit from the `historical_data`.

If round-specific data like qualifying position or practice ranks are not
provided in the `drivers` data frame, the function will estimate or impute
them based on historical averages or sensible defaults. For example, if `grid`
is missing, it will be estimated based on the driver's historical average
grid position.

The features generated include driver experience, failure rates, average
finishing positions, constructor averages, and circuit-specific performance
metrics.
}
\examples{
\dontrun{
new_data <- generate_new_data(season = 2024, round = 1)
}
}
