% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_grids}
\alias{get_grids}
\title{Get Quali position, starting grid, and finishing position}
\usage{
get_grids(season, race, session)
}
\arguments{
\item{season}{season}

\item{race}{race}

\item{session}{Usually 'R', but also available is 'S'}
}
\value{
data frame
}
\description{
Get Quali position, starting grid, and finishing position
}
