% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_laps}
\alias{get_laps}
\title{Get Weekend's laps}
\usage{
get_laps(season, race)
}
\arguments{
\item{season}{season, from 2018 onward}

\item{race}{race, compatable with fastf1 race selector}
}
\value{
a data.frame with all lap information
}
\description{
Get all laps (timing) from a weekend and return as one data.frame
}
