% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{get_race_name}
\alias{get_race_name}
\title{Get Race Name}
\usage{
get_race_name(season, round)
}
\arguments{
\item{season}{The season year}

\item{round}{The round number}
}
\value{
A string with the race name, e.g. "Bahrain Grand Prix"
}
\description{
Get Race Name
}
