% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_schedule}
\alias{get_schedule}
\title{Get the schedule}
\usage{
get_schedule(save_data = FALSE)
}
\arguments{
\item{save_data}{whether or not to update the saved schedule object}
}
\value{
Schedule data.frame
}
\description{
Get the schedule
}
