% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_season_data}
\alias{get_season_data}
\title{Get Season Data}
\usage{
get_season_data(season, force = FALSE)
}
\arguments{
\item{season}{a season, from 2018 onward}

\item{force}{whether to force reloading data}
}
\value{
None, writes to file
}
\description{
Get a whole season's data (up to today's date, if current data). Saves data to
options('f1predicter.cache') directory
}
