% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_weekend_data}
\alias{get_weekend_data}
\title{Get weekend's Data}
\usage{
get_weekend_data(season, round, force = FALSE)
}
\arguments{
\item{season}{season from 2018 onward}

\item{round}{round number}

\item{force}{whether to force getting new data}
}
\value{
a list of data frames
}
\description{
Downloads and saves all data for a round weekend
}
