% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getWeather}
\alias{getWeather}
\title{getRaceWeather}
\usage{
getWeather(race_url)
}
\arguments{
\item{race_url}{a wikipedia url for a grand prix (in english)}
}
\value{
a weather, as character (one of `warm`, `cold`, `dry`, `wet`, or `cloudy`)
}
\description{
given a race url from ergast (to wikipedia) get the weather
}
