% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_all_data}
\alias{load_all_data}
\title{Load All Data}
\usage{
load_all_data()
}
\value{
a list of data.frames
}
\description{
Loads all data previously saved to cache at options('f1predicter.cache') location
}
