% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ensemble.R
\name{load_ensemble_model}
\alias{load_ensemble_model}
\title{Load a Stacked Ensemble Model}
\usage{
load_ensemble_model(model_name)
}
\arguments{
\item{model_name}{The original name of the ensemble model (e.g., "Quali Position Ensemble").}
}
\value{
A loaded `model_stack` object.
}
\description{
Loads a previously saved `model_stack` object.
}
