% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_reg.R
\name{load_models}
\alias{load_models}
\title{Load Models}
\usage{
load_models(model_type, model_timing, engine = "ranger")
}
\arguments{
\item{model_type}{The type of models to load, either `"quali"` or `"results"`.}

\item{model_timing}{The timing context for the models, one of `"early"`,
`"late"`, or (for "results" models only) `"after_quali"`.}

\item{engine}{The engine used for the model, e.g., `"ranger"`, `"glmnet"`, `"nnet"`, `"kernlab"`, or `"kknn"`.}
}
\value{
A list of  model objects, ready for prediction.
}
\description{
Loads a list of models from a standardized file path. The path is
constructed from the base directory set in `getOption("f1predicter.models")`,
the model type, and timing.
}
