% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_class_glmnet.R
\name{model_quali_late_glm}
\alias{model_quali_late_glm}
\title{Model Qualification Results Using GLM}
\usage{
model_quali_late_glm(data = clean_data())
}
\arguments{
\item{data}{should be clean_data() but modified data of the same structure can be used.}
}
\description{
Build a model for quali using GLM

Use `model_quali_early_glm` or `model_quali_late_glm` depending on if you have practice data included for that weekend.
}
