% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_class_xgb.R
\name{model_quali_late_xgb}
\alias{model_quali_late_xgb}
\title{Model Qualification Results Using XGB}
\usage{
model_quali_late_xgb(data = clean_data())
}
\arguments{
\item{data}{should be clean_data() but modified data of the same structure can be used.}
}
\description{
Build a model for quali using XGBoost.

Use `model_quali_early_xgb` or `model_quali_late_xgb` depending on if you have practice data included for that weekend.
}
