% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_vector}
\alias{normalize_vector}
\title{Normalize a Numeric Vector}
\usage{
normalize_vector(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector of the same length as `x` where the elements sum to 1.
}
\description{
This function takes a numeric vector and scales its values so that they sum
to 1. `NA` values are treated as 0.
}
\details{
If the sum of the vector is 0 (e.g., it contains only zeros and `NA`s),
the function will return a vector of uniform values, where each element is
`1 / length(x)`. Otherwise, each element is divided by the total sum of the
vector after `NA`s have been converted to 0.
}
\examples{
normalize_vector(c(1, 2, 3, 4))
normalize_vector(c(10, 10))
normalize_vector(c(1, 2, NA, 3))
normalize_vector(c(0, 0, 0))
normalize_vector(c(NA, NA))
}
