% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{post_race_odds}
\alias{post_race_odds}
\title{Post Race Odds to Bluesky}
\usage{
post_race_odds(predictions)
}
\arguments{
\item{predictions}{A data frame of predictions from `predict_round()`.}
}
\description{
Post Race Odds to Bluesky
}
