% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_finish}
\alias{predict_finish}
\title{Predict Race Finish (Not DNF)}
\usage{
predict_finish(new_data = generate_new_data(), finish_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{finish_model}{A `workflow` object for predicting finishing the race.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `finish_odd`.
}
\description{
Predict Race Finish (Not DNF)
}
\keyword{internal}
