% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_podium}
\alias{predict_podium}
\title{Predict Podium Finish}
\usage{
predict_podium(new_data = generate_next_race_data(), podium_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{podium_model}{A `workflow` object for predicting a podium finish.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `podium_odd`.
}
\description{
Predict Podium Finish
}
\keyword{internal}
