% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_position}
\alias{predict_position}
\title{Predict Finishing Position}
\usage{
predict_position(new_data = generate_new_data(), position_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{position_model}{A `workflow` object for predicting the finishing position.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `likely_position`.
}
\description{
Predict Finishing Position
}
\keyword{internal}
