% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_position_class}
\alias{predict_position_class}
\title{Predict Finishing Position (Classification)}
\usage{
predict_position_class(
  new_data = generate_next_race_data(),
  position_class_model
)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{position_class_model}{A list containing the `polr` fit and recipe.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `likely_position_class`.
}
\description{
Predict Finishing Position (Classification)
}
\keyword{internal}
