% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_quali_pole}
\alias{predict_quali_pole}
\title{Predict Pole Position}
\usage{
predict_quali_pole(new_data = generate_next_race_data(), quali_pole_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{quali_pole_model}{A `workflow` object for predicting pole position,
such as `model_quali_early()$quali_pole`.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and `pole_odd` (the
  predicted probability of getting pole position).
}
\description{
Predicts the probability of each driver achieving pole position for an upcoming
race.
}
\details{
This function takes a `workflow` object (trained for pole position prediction)
and a data frame of features for the upcoming race. The user can control
whether to make an "early" (pre-practice) or "late" (post-practice)
prediction by passing the appropriate model object from `model_quali_early()`
or `model_quali_late()`.
}
