% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_quali_pos}
\alias{predict_quali_pos}
\title{Predict Qualifying Position}
\usage{
predict_quali_pos(new_data = generate_next_race_data(), quali_pos_model)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{quali_pos_model}{A `workflow` object for predicting qualifying position,
such as `model_quali_early()$quali_pos`.}
}
\value{
A tibble with `driver_id`, `round`, `season`, and
  `likely_quali_position`.
}
\description{
Predicts the likely qualifying position for each driver in an upcoming race.
}
\details{
This function takes a `workflow` object (trained for qualifying position
prediction) and a data frame of features for the upcoming race. The user can
control whether to make an "early" (pre-practice) or "late" (post-practice)
prediction by passing the appropriate model object from `model_quali_early()`
or `model_quali_late()`.
}
