% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_quali_round}
\alias{predict_quali_round}
\title{Predict Qualifying Results for a Round}
\usage{
predict_quali_round(
  new_data = generate_next_race_data(),
  quali_pole_model,
  quali_pos_model
)
}
\arguments{
\item{new_data}{A data frame of new data, typically from `generate_new_data()`.}

\item{quali_pole_model}{A `workflow` object for predicting pole position.}

\item{quali_pos_model}{A `workflow` object for predicting qualifying position.}
}
\value{
A tibble with predictions for pole probability and qualifying position
  for each driver.
}
\description{
A wrapper function to predict both pole position probability and likely
qualifying position for a given round.
}
\details{
This function combines the outputs of `predict_quali_pole()` and
`predict_quali_pos()` into a single tibble.
}
