% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_reg.R
\name{prepare_and_split_data}
\alias{prepare_and_split_data}
\title{Prepare and Split Data for Modeling}
\usage{
prepare_and_split_data(data, columns = NULL, prop = 4/5, group = "round_id")
}
\arguments{
\item{data}{The input data frame.}

\item{columns}{A character vector of column names to select. If `NULL`,
all columns in `data` are used.}

\item{prop}{The proportion of data to be in the training set.}

\item{group}{The variable to group by for the split.}
}
\value{
A list containing the data split object, training data,
  testing data, and cross-validation folds.
}
\description{
This helper function preprocesses the data by selecting columns,
converting character columns to factors, and then splits the data into
training/testing sets and creates cross-validation folds.
}
