% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schedule}
\alias{schedule}
\title{Schedule Data}
\description{
This contains the schedule data. Can regenerate by calling `get_schedule()`
}
\references{
derived from ergast data
}
\keyword{data}
