% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_reg.R
\name{train_results_models}
\alias{train_results_models}
\title{Train Race Result Models}
\usage{
train_results_models(data, scenario, engine = "xgboost")
}
\arguments{
\item{data}{A data frame containing the modeling data.}

\item{scenario}{A character string specifying the modeling context. One of
"early" (pre-practice), "late" (post-practice), or "after_quali".}

\item{engine}{A character string specifying the model engine. One of "xgboost"
(default) or "glmnet".}
}
\value{
A list containing five fitted `workflow` objects.
}
\description{
This function trains a suite of five classification models to predict race
results based on the specified scenario.
}
