% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{win}
\alias{win}
\title{Win metric}
\usage{
win(data, ...)
}
\arguments{
\item{data}{data}

\item{...}{other elements}
}
\value{
a numeric value between 0 and 1 (higher is better)
}
\description{
Subsets the normal accuracy metric to only count the target position. For example, picking all 0 for win chance would
normally give a 95% accuracy (1 in 20 is wrong - 1 in 20 wins). This function only looks at the accuracy relative
to the truth == 1 case (i.e. correctly or not pick the winner). Works for win or for podium or points models
}
\seealso{
[win_vec()] for vector version
}
