% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{win_vec}
\alias{win_vec}
\title{Win metric}
\usage{
win_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{truth}{truth value (0 or 1)}

\item{estimate}{estimates (note that they're limited to 0..1 and are rounded in the function to either 0 or 1)}

\item{case_weights}{ignored}
}
\value{
a numeric value between 0 and 1 (higher is better)
}
\description{
Subsets the normal accuracy metric to only count the target position. For example, picking all 0 for win chance would
normally give a 95% accuracy (1 in 20 is wrong - 1 in 20 wins). This function only looks at the accuracy relative
to the truth == 1 case (i.e. correctly or not pick the winner). Works for win or for podium or points models
}
\seealso{
[win()] for data.frame version
}
