% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(input = load_all_data())
}
\arguments{
\item{input}{A list of raw data frames, typically from `load_all_data()`.}
}
\value{
A single, cleaned data frame.
}
\description{
This function serves as a wrapper to load all raw data and process it through
a series of cleaning and feature engineering steps, preparing it for modeling.
}
