% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{clean_data}
\alias{clean_data}
\title{Clean Data}
\usage{
clean_data(input = load_all_data())
}
\arguments{
\item{input}{take an input data (or load fresh) and process it for modelling}
}
\value{
a data frame
}
\description{
Clean Data
}
