% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_ensemble.R
\name{get_hyperparameters}
\alias{get_hyperparameters}
\title{Get Optimal Model hyperparameters (2025-09)}
\usage{
get_hyperparameters(model = "quali", timing = "early")
}
\arguments{
\item{model}{What type of model to load hyperparameters for}

\item{timing}{Early, late, or after-quali?}
}
\value{
a list of list of hyperparameters by engine.
}
\description{
Model hyperparameters were determined in 2025-09
and are conveniently returned by this function
}
