% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_reg.R
\name{model_results_late}
\alias{model_results_late}
\title{Train Race Result Models (Post-Practice)}
\usage{
model_results_late(data = clean_data())
}
\arguments{
\item{data}{A data frame containing the modeling data. Defaults to `clean_data()`.}
}
\value{
A list containing five fitted `workflow` objects for `win`, `podium`,
  `t10`, `finish`, and `position`.
}
\description{
This function trains a suite of five classification models to predict race
results using data available *after* practice sessions but *before*
qualifying.

It produces the same models as `model_results_after_quali`, but uses a
smaller set of predictors.
}
\details{
The function filters data for seasons from 2018 onwards. All five models are
built using the `xgboost` engine, with hyperparameters tuned via a Latin
hypercube grid search.

The four binary models (`win`, `podium`, `t10`, `finish`) are optimized on
the `mn_log_loss` metric. The `position` regression model is optimized
based on the `rmse` (Root Mean Squared Error) metric.
}
