% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{post_skeet_predictions}
\alias{post_skeet_predictions}
\title{Post Predictions to Bluesky}
\usage{
post_skeet_predictions(texts, image = NULL, image_alt = NULL, tags = NULL)
}
\arguments{
\item{texts}{A character vector where each element is the text for a post in a thread.}

\item{image}{file path to a graphic image, if to be added}

\item{image_alt}{Alternative text for the image, for accessibility.}

\item{tags}{A character vector of tags to apply to the post.}
}
\value{
Invisibly returns the response from the Bluesky API, or NULL on failure.
}
\description{
This function takes a status string and posts it to Bluesky.
It requires credentials to be set as environment variables for `atrrr`.
See the atrrr documentation for details.
}
