% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_reg.R
\name{save_models}
\alias{save_models}
\title{Save Models}
\usage{
save_models(model_list, model_timing)
}
\arguments{
\item{model_list}{A named list of `last_fit` objects from `tune`. The names
(e.g., "quali_pole", "win") are used to infer the model type.}

\item{model_timing}{The timing context for the models, one of `"early"`,
`"late"`, or (for "results" models only) `"after_quali"`.}
}
\value{
Invisibly returns the full `file_path` where models were saved.
}
\description{
This function takes a list of trained `last_fit` model object
and saves the resulting list to a standardized file
path. The `model_type` ("quali" or "results") is automatically inferred from
the names of the models in the list (e.g., "quali_pole", "win").
}
\details{
The path is constructed from the base directory set in
`getOption("f1predicter.models")` and the specified model timing.
}
